16HW-переробив фікстуру
